more (mostly esoteric) opts.
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@66 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ media/lzwreader.h
@@ -46,7 +46,7 @@ public:
 	int get_byte();
 	BYTE *NextLine();
 
-	short init_exp(short size);
+	short init_exp(int size);
 	int get_next_code();
 	short Decoder();
